Change VFS directory function to own header
