+++ addons/library.kodi.addon/addon.api2/internal/libKODI_addon_VFSUtilsHead.h
